+++ scripts/8000_postinst.sh
@@ -1,5 +1,5 @@
 # run postinst
-if [ -f "$PDK_WORKSPACE/$POSTINST" ]; then
+if [ ! -z "$POSTINST" ] && [ -f "$PDK_WORKSPACE/$POSTINST" ]; then
   info "running postinst $POSTINST"
   cp "$PDK_WORKSPACE/$POSTINST" "$ROOTFS/tmp/postinst"
   chmod +x "$ROOTFS/tmp/postinst"
